refactor: split delivery target runtime seams
