+++ src/cron/isolated-agent/delivery-target.test.ts
@@ -4,6 +4,10 @@ import type { OpenClawConfig } from "../../config/config.js";
 import { resetPluginRuntimeStateForTest, setActivePluginRegistry } from "../../plugins/runtime.js";
 import { createOutboundTestPlugin, createTestRegistry } from "../../test-utils/channel-plugins.js";
 
+const whatsappAccountMocks = vi.hoisted(() => ({
+  resolveWhatsAppAccount: vi.fn<() => { allowFrom: string[] }>(() => ({ allowFrom: [] })),
+}));
+
 vi.mock("../../config/sessions/main-session.js", () => ({
   resolveAgentMainSessionKey: vi.fn().mockReturnValue("agent:test:main"),
 }));
@@ -16,7 +20,7 @@ vi.mock("../../config/sessions/store-load.js", () => ({
   loadSessionStore: vi.fn().mockReturnValue({}),
 }));
 
-vi.mock("../../infra/outbound/channel-selection.js", () => ({
+vi.mock("../../infra/outbound/channel-selection.runtime.js", () => ({
   resolveMessageChannelSelection: vi
     .fn()
     .mockResolvedValue({ channel: "telegram", configured: ["telegram"] }),
@@ -30,18 +34,28 @@ vi.mock("../../pairing/pairing-store.js", () => ({
   readChannelAllowFromStoreSync: vi.fn(() => []),
 }));
 
+vi.mock("../../plugin-sdk/whatsapp-surface.js", () => ({
+  resolveWhatsAppAccount: whatsappAccountMocks.resolveWhatsAppAccount,
+}));
+
+vi.mock("../../infra/outbound/targets.runtime.js", () => ({
+  resolveOutboundTarget: vi.fn(),
+}));
 const mockedModuleIds = [
   "../../config/sessions/main-session.js",
   "../../config/sessions/paths.js",
   "../../config/sessions/store-load.js",
-  "../../infra/outbound/channel-selection.js",
+  "../../infra/outbound/channel-selection.runtime.js",
+  "../../infra/outbound/targets.runtime.js",
   "../../infra/outbound/target-resolver.js",
   "../../pairing/pairing-store.js",
+  "../../plugin-sdk/whatsapp-surface.js",
 ];
 
 import { loadSessionStore } from "../../config/sessions/store-load.js";
-import { resolveMessageChannelSelection } from "../../infra/outbound/channel-selection.js";
+import { resolveMessageChannelSelection } from "../../infra/outbound/channel-selection.runtime.js";
 import { maybeResolveIdLikeTarget } from "../../infra/outbound/target-resolver.js";
+import { resolveOutboundTarget } from "../../infra/outbound/targets.runtime.js";
 import { readChannelAllowFromStoreSync } from "../../pairing/pairing-store.js";
 import { resolveDeliveryTarget } from "./delivery-target.js";
 
@@ -82,6 +96,7 @@ function createAllowlistAwareStubOutbound(label: string): ChannelOutboundAdapter
 
 beforeEach(() => {
   resetPluginRuntimeStateForTest();
+  vi.mocked(resolveOutboundTarget).mockReset();
   setActivePluginRegistry(
     createTestRegistry([
       {
@@ -275,6 +290,42 @@ describe("resolveDeliveryTarget", () => {
       expect.objectContaining({
         channel: "telegram",
         input: "123456789",
+      }),
+    );
+  });
+
+  it("falls back to the runtime target resolver when the channel plugin is not already loaded", async () => {
+    setMainSessionEntry(undefined);
+    setActivePluginRegistry(
+      createTestRegistry([
+        {
+          pluginId: "whatsapp",
+          plugin: createOutboundTestPlugin({
+            id: "whatsapp",
+            outbound: createStubOutbound("WhatsApp"),
+          }),
+          source: "test",
+        },
+      ]),
+    );
+    vi.mocked(resolveOutboundTarget).mockReturnValueOnce({ ok: true, to: "123456" });
+
+    const result = await resolveDeliveryTarget(makeCfg({ bindings: [] }), AGENT_ID, {
+      channel: "telegram",
+      to: "123456",
+    });
+
+    expect(result).toEqual(
+      expect.objectContaining({
+        ok: true,
+        channel: "telegram",
+        to: "123456",
+      }),
+    );
+    expect(resolveOutboundTarget).toHaveBeenCalledWith(
+      expect.objectContaining({
+        channel: "telegram",
+        to: "123456",
       }),
     );
   });
