+++ src/infra/outbound/channel-selection.runtime.ts
@@ -1,0 +1,1 @@
+export { resolveMessageChannelSelection } from "./channel-selection.js";
