+++ src/infra/outbound/targets.runtime.ts
@@ -1,0 +1,1 @@
+export { resolveOutboundTarget } from "./targets.js";
