+++ src/infra/outbound/targets.ts
@@ -1,10 +1,5 @@
 import { mapAllowFromEntries } from "openclaw/plugin-sdk/channel-config-helpers";
 import { normalizeChatType, type ChatType } from "../../channels/chat-type.js";
-import {
-  comparableChannelTargetsShareRoute,
-  parseExplicitTargetForChannel,
-  resolveComparableTargetForChannel,
-} from "../../channels/plugins/target-parsing.js";
 import type { ChannelOutboundTargetMode } from "../../channels/plugins/types.js";
 import { formatCliCommand } from "../../cli/command-format.js";
 import type { OpenClawConfig } from "../../config/config.js";
@@ -23,7 +18,6 @@ import type {
 import {
   INTERNAL_MESSAGE_CHANNEL,
   isDeliverableMessageChannel,
-  normalizeMessageChannel,
 } from "../../utils/message-channel.js";
 import {
   normalizeDeliverableOutboundChannel,
@@ -53,175 +47,8 @@ export type HeartbeatSenderContext = {
 
 export type OutboundTargetResolution = { ok: true; to: string } | { ok: false; error: Error };
 
-export type SessionDeliveryTarget = {
-  channel?: DeliverableMessageChannel;
-  to?: string;
-  accountId?: string;
-  threadId?: string | number;
-  /** Whether threadId came from an explicit source (config/param/:topic: parsing) vs session history. */
-  threadIdExplicit?: boolean;
-  mode: ChannelOutboundTargetMode;
-  lastChannel?: DeliverableMessageChannel;
-  lastTo?: string;
-  lastAccountId?: string;
-  lastThreadId?: string | number;
-};
-
-function parseExplicitTargetWithPlugin(params: {
-  channel?: DeliverableMessageChannel;
-  fallbackChannel?: DeliverableMessageChannel;
-  raw?: string;
-}) {
-  const raw = params.raw?.trim();
-  if (!raw) {
-    return null;
-  }
-  const provider = params.channel ?? params.fallbackChannel;
-  if (!provider) {
-    return null;
-  }
-  return parseExplicitTargetForChannel(provider, raw);
-}
-
-export function resolveSessionDeliveryTarget(params: {
-  entry?: SessionEntry;
-  requestedChannel?: GatewayMessageChannel | "last";
-  explicitTo?: string;
-  explicitThreadId?: string | number;
-  fallbackChannel?: DeliverableMessageChannel;
-  allowMismatchedLastTo?: boolean;
-  mode?: ChannelOutboundTargetMode;
-  /**
-   * When set, this overrides the session-level `lastChannel` for "last"
-   * resolution.  This prevents cross-channel reply routing when multiple
-   * channels share the same session (dmScope = "main") and an inbound
-   * message from a different channel updates `lastChannel` while an agent
-   * turn is still in flight.
-   *
-   * Callers should set this to the channel that originated the current
-   * agent turn so the reply always routes back to the correct channel.
-   *
-   * @see https://github.com/openclaw/openclaw/issues/24152
-   */
-  turnSourceChannel?: DeliverableMessageChannel;
-  /** Turn-source `to` — paired with `turnSourceChannel`. */
-  turnSourceTo?: string;
-  /** Turn-source `accountId` — paired with `turnSourceChannel`. */
-  turnSourceAccountId?: string;
-  /** Turn-source `threadId` — paired with `turnSourceChannel`. */
-  turnSourceThreadId?: string | number;
-}): SessionDeliveryTarget {
-  const context = deliveryContextFromSession(params.entry);
-  const sessionLastChannel =
-    context?.channel && isDeliverableMessageChannel(context.channel) ? context.channel : undefined;
-  const parsedSessionTarget = sessionLastChannel
-    ? resolveComparableTargetForChannel({
-        channel: sessionLastChannel,
-        rawTarget: context?.to,
-        fallbackThreadId: context?.threadId,
-      })
-    : null;
-
-  // When a turn-source channel is provided, use only turn-scoped metadata.
-  // Falling back to mutable session fields would re-introduce routing races.
-  const hasTurnSourceChannel = params.turnSourceChannel != null;
-  const parsedTurnSourceTarget =
-    hasTurnSourceChannel && params.turnSourceChannel
-      ? resolveComparableTargetForChannel({
-          channel: params.turnSourceChannel,
-          rawTarget: params.turnSourceTo,
-          fallbackThreadId: params.turnSourceThreadId,
-        })
-      : null;
-  const hasTurnSourceThreadId = parsedTurnSourceTarget?.threadId != null;
-  const lastChannel = hasTurnSourceChannel ? params.turnSourceChannel : sessionLastChannel;
-  const lastTo = hasTurnSourceChannel ? params.turnSourceTo : context?.to;
-  const lastAccountId = hasTurnSourceChannel ? params.turnSourceAccountId : context?.accountId;
-  // Fall back to the session's stored threadId only when the turn-source channel AND destination
-  // match the session context. This avoids mixing a turn-scoped `to` with a stale session-scoped
-  // threadId from a different chat/topic in shared-session scenarios.
-  const turnToMatchesSession =
-    !params.turnSourceTo ||
-    !context?.to ||
-    (params.turnSourceChannel === sessionLastChannel &&
-      comparableChannelTargetsShareRoute({
-        left: parsedTurnSourceTarget,
-        right: parsedSessionTarget,
-      }));
-  const lastThreadId = hasTurnSourceThreadId
-    ? parsedTurnSourceTarget?.threadId
-    : hasTurnSourceChannel &&
-        (params.turnSourceChannel !== sessionLastChannel || !turnToMatchesSession)
-      ? undefined
-      : parsedSessionTarget?.threadId;
-
-  const rawRequested = params.requestedChannel ?? "last";
-  const requested = rawRequested === "last" ? "last" : normalizeMessageChannel(rawRequested);
-  const requestedChannel =
-    requested === "last"
-      ? "last"
-      : requested && isDeliverableMessageChannel(requested)
-        ? requested
-        : undefined;
-
-  const rawExplicitTo =
-    typeof params.explicitTo === "string" && params.explicitTo.trim()
-      ? params.explicitTo.trim()
-      : undefined;
-
-  let channel = requestedChannel === "last" ? lastChannel : requestedChannel;
-  if (!channel && params.fallbackChannel && isDeliverableMessageChannel(params.fallbackChannel)) {
-    channel = params.fallbackChannel;
-  }
-
-  let explicitTo = rawExplicitTo;
-  const parsedExplicitTarget = parseExplicitTargetWithPlugin({
-    channel,
-    fallbackChannel: !channel ? lastChannel : undefined,
-    raw: rawExplicitTo,
-  });
-  if (parsedExplicitTarget?.to) {
-    explicitTo = parsedExplicitTarget.to;
-  }
-  const explicitThreadId =
-    params.explicitThreadId != null && params.explicitThreadId !== ""
-      ? params.explicitThreadId
-      : parsedExplicitTarget?.threadId;
-
-  let to = explicitTo;
-  if (!to && lastTo) {
-    if (channel && channel === lastChannel) {
-      to = lastTo;
-    } else if (params.allowMismatchedLastTo) {
-      to = lastTo;
-    }
-  }
-
-  const mode = params.mode ?? (explicitTo ? "explicit" : "implicit");
-  const accountId = channel && channel === lastChannel ? lastAccountId : undefined;
-  const threadId =
-    channel && channel === lastChannel
-      ? mode === "heartbeat"
-        ? hasTurnSourceThreadId
-          ? params.turnSourceThreadId
-          : undefined
-        : lastThreadId
-      : undefined;
-
-  const resolvedThreadId = explicitThreadId ?? threadId;
-  return {
-    channel,
-    to,
-    accountId,
-    threadId: resolvedThreadId,
-    threadIdExplicit: resolvedThreadId != null && explicitThreadId != null,
-    mode,
-    lastChannel,
-    lastTo,
-    lastAccountId,
-    lastThreadId,
-  };
-}
+export { resolveSessionDeliveryTarget, type SessionDeliveryTarget } from "./targets-session.js";
+import { resolveSessionDeliveryTarget } from "./targets-session.js";
 
 // Channel docking: prefer plugin.outbound.resolveTarget + allowFrom to normalize destinations.
 export function resolveOutboundTarget(params: {
